Changed plain text printing to be Tj.

+++ src/document.hpp
@@ -165,7 +165,7 @@ struct DelayedPage {
 
 struct SubsetGlyph {
     FontSubset ss;
-    uint8_t glyph_id;
+    uint16_t glyph_id;
 };
 
 struct FontThingy {
